fix: create api and hook for search modal
