+++ src/components/header/components/searchDropdown/CategoryTabs/CategoryCard/CategoryCard.styles.ts
@@ -1,8 +1,7 @@
-import React from "react";
 import { Card } from "antd";
 import styled from "styled-components";
 
-const StyledCard = styled(Card)`
+export const StyledCard = styled(Card)`
   width: 150px;
   height: 100px;
   margin: 8px;
@@ -23,7 +22,7 @@ const StyledCard = styled(Card)`
   }
 `;
 
-const ImageContainer = styled.div`
+export const ImageContainer = styled.div`
   width: 100%;
   height: 100%;
   position: absolute;
@@ -31,13 +30,13 @@ const ImageContainer = styled.div`
   left: 0;
 `;
 
-const Image = styled.img`
+export const Image = styled.img`
   width: 100%;
   height: 100%;
   object-fit: cover;
 `;
 
-const TitleOverlay = styled.span`
+export const TitleOverlay = styled.span`
   position: absolute;
   top: 8px;
   left: 8px;
@@ -49,19 +48,3 @@ const TitleOverlay = styled.span`
   text-shadow: 1px 1px 4px rgba(0, 0, 0, 0.8);
   z-index: 2;
 `;
-
-interface CategoryCardProps {
-  title: string;
-  image: string;
-}
-
-const CategoryCard: React.FC<CategoryCardProps> = ({ title, image }) => (
-  <StyledCard>
-    <ImageContainer>
-      <Image src={image} alt={title} />
-    </ImageContainer>
-    <TitleOverlay>{title}</TitleOverlay>
-  </StyledCard>
-);
-
-export default CategoryCard;
