+++ src/locales/en/translation.json
@@ -1,4 +1,5 @@
 {
+  "monthNames": ["Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec"],
   "searchBar":{
     "browseByCategory": "Browse by Category",
     "browseByLocation": "Browse by City",
