+++ src/locales/vn/translation.json
@@ -1,4 +1,5 @@
 {
+  "monthNames": ["Th1", "Th2", "Th3", "Th4", "Th5", "Th6", "Th7", "Th8", "Th9", "Th10", "Th11", "Th12"],
   "searchBar": {
     "browseByCategory": "Khám phá theo Thể loại",
     "browseByLocation": "Khám phá theo Thành phố",
